# openplatform-api-service/docs/openapi/developer-api.yaml
openapi: 3.0.3
info:
  title: Cregis Custody OpenPlatform Developer API
  description: |
    Cregis Custody 开放平台开发者 API 文档

    ## 认证方式
    所有需要认证的接口需要在 Header 中携带 Authorization: Bearer {accessToken}

    ## 错误响应格式
    ```json
    {
      "code": 0,
      "message": "success",
      "data": { ... },
      "trace_id": "string"
    }
    ```

    ## 错误码说明
    - **0**: 成功
    - **400xx**: 请求参数错误
    - **401xx**: 认证错误（令牌无效/过期）
    - **403xx**: 权限错误（无访问权限）
    - **404xx**: 资源不存在
    - **409xx**: 业务冲突（如重复注册）
    - **500xx**: 服务器内部错误
  version: 1.0.0
  contact:
    name: Cregis Tech
    email: support@cregis.io
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发环境
  - url: https://api.cregis.io/v1
    description: 生产环境
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: 用户注册
      description: |
        注册新的开发者账号，需要填写企业信息和UBO（最终受益人）信息。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterParams'
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterParams'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 邮箱已被注册
  /auth/login:
    post:
      tags:
        - Authentication
      summary: 用户登录
      description: 使用邮箱密码登录系统
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: 用户邮箱
                  example: developer@example.com
                password:
                  type: string
                  format: password
                  description: 密码
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 邮箱或密码错误
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: 用户登出
      description: 清除当前会话
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: 忘记密码
      description: 发送密码重置邮件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: 注册邮箱
      responses:
        '200':
          description: 重置邮件已发送
        '404':
          description: 邮箱未注册
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: 重置密码
      description: 使用重置令牌设置新密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  description: 重置令牌
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: 新密码（至少8位）
      responses:
        '200':
          description: 密码重置成功
        '400':
          description: 令牌无效或已过期
...
  /applications:
    get:
      tags:
        - Applications
      summary: 获取应用列表
      description: 获取当前用户的所有应用
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
        - in: query
          name: keyword
          schema:
            type: string
          description: 搜索关键词
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse_Application'

    post:
      tags:
        - Applications
      summary: 创建应用
      description: 创建一个新的应用
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                  description: 应用名称
                description:
                  type: string
                  maxLength: 500
                  description: 应用描述
                callbackUrl:
                  type: string
                  format: uri
                  description: 回调地址
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
  /applications/{id}:
    get:
      tags:
        - Applications
      summary: 获取应用详情
      description: 获取指定应用的详细信息
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: 应用ID
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetail'

    put:
      tags:
        - Applications
      summary: 更新应用
      description: 更新应用信息
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: 应用ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                callbackUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: 更新成功

    delete:
      tags:
        - Applications
      summary: 删除应用
      description: 删除指定应用
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: 应用ID
      responses:
        '200':
          description: 删除成功
  /applications/{id}/regenerate-secret:
    post:
      tags:
        - Applications
      summary: 重新生成AppSecret
      description: 重新生成应用的 AppSecret，原 Secret 将立即失效
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: 应用ID
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  app_secret:
                    type: string
                    description: 新的 AppSecret

  # ==================== Billing ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: 错误码
        message:
          type: string
          description: 错误信息
        data:
          type: object
          nullable: true
        trace_id:
          type: string

    # ==================== Authentication ====================
    RegisterParams:
      type: object
      required:
        - email
        - password
        - companyName
        - creditCode
        - industry
        - ubos
        - orgType
      properties:
        email:
          type: string
          format: email
          description: 注册邮箱
        password:
          type: string
          format: password
          minLength: 8
          description: 密码（至少8位）
        companyName:
          type: string
          minLength: 2
          maxLength: 100
          description: 企业名称
        creditCode:
          type: string
          description: 统一社会信用代码
        website:
          type: string
          format: uri
          description: 企业网站
        industry:
          type: string
          description: 所属行业
        ubos:
          type: array
          minItems: 1
          description: 最终受益人信息
          items:
            type: object
            required:
              - name
              - idType
              - idNumber
              - nationality
              - phone
            properties:
              name:
                type: string
                description: 姓名
              idType:
                type: string
                enum: [passport, national_id]
                description: 证件类型
              idNumber:
                type: string
                description: 证件号码
              nationality:
                type: string
                description: 国籍
              phone:
                type: string
                description: 电话
        orgType:
          type: string
          description: 组织类型
        establishmentDate:
          type: string
          format: date
          description: 成立日期
        registeredAddress:
          type: string
          description: 注册地址
        businessScope:
          type: string
          description: 经营范围
...
    AuthResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            accessToken:
              type: string
              description: 访问令牌
            refreshToken:
              type: string
              description: 刷新令牌
            user:
              $ref: '#/components/schemas/UserProfile'
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: 用户ID
        email:
          type: string
          format: email
          description: 邮箱
        companyName:
          type: string
          description: 企业名称
        status:
          type: string
          enum: [pending, approved, rejected]
          description: 账户状态
        kybStatus:
          type: string
          enum: [pending, approved, rejected]
          description: KYB审核状态
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间

    # ==================== Applications ====================
    Application:
      type: object
      properties:
        id:
          type: string
          description: 应用ID
        name:
          type: string
          description: 应用名称
        appId:
          type: string
          description: 应用标识
        appSecret:
          type: string
          description: 应用密钥（仅创建时返回）
        status:
          type: string
          enum: [pending_review, active, inactive, suspended]
          description: 应用状态
        description:
          type: string
          description: 应用描述
        callbackUrl:
          type: string
          description: 回调地址
        apiUsage:
          $ref: '#/components/schemas/ApplicationApiUsage'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicationDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/Application'
        - type: object
          properties:
            credentials:
              type: object
              properties:
                appId:
                  type: string
                appSecret:
                  type: string
                  format: password
              description: 凭证信息（创建或重新生成时返回）
            security:
              type: object
              properties:
                ipWhitelist:
                  type: array
                  items:
                    type: string
                  description: IP白名单
                rateLimit:
                  type: string
                  description: 速率限制等级
    ApplicationApiUsage:
      type: object
      properties:
        totalCalls:
          type: integer
          description: 总调用次数
        last30Days:
          type: integer
          description: 近30天调用次数
        successRate:
          type: number
          description: 成功率
...
    ListResponse_Application:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # ==================== Billing ====================
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 40001
            message: "Invalid request parameters"
            trace_id: "abc123"
...